ai-chatbot/backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.7
redis==5.0.0
celery==5.3.1
requests==2.31.0
python-dotenv==1.0.0
django-cors-headers==4.3.0
djangorestframework-simplejwt==5.3.0
ollama==0.1.7
chromadb==0.4.10
langchain==0.1.0
gunicorn==21.2.0
beautifulsoup4==4.12.2
lxml==4.9.3
whitenoise==6.6.0

# Embedding model - yaxshiroq sifat uchun
sentence-transformers>=2.2.2
torch>=2.1.0

# Document processing dependencies
PyPDF2==3.0.1
python-docx==1.1.0

# NEW: Hybrid Search (BM25)
rank-bm25==0.2.2

# NEW: Text similarity for validation
scikit-learn>=1.3.0
langdetect==1.0.9

# NEW: Error Monitoring (Sentry)
sentry-sdk>=1.40.0